Fix variable parser fallacies
The `extract_var_from_token` used `ft_split` that splits at a given
delimeter (equal sign).
If there were no characters after the equal sign `ft_split` did not
return an empty string as variable value. This commit fixes this
issue, and uses `ft_substr` instead. It takes a substring from the
start of the string up to the equal sign as name, and from the equal sign
up the end of the string as value.