Fix segfaulting `unset`
The issue presents itself when a variable is defined in the shell scope but not
in the environment, i.e. not exported. When the variable is unset, a pointer is
assigned to point to the node to be removed. However this pointer was used for
both the shell and the environment list, and hence once the node was removed
form the shell scope, the pointer tried to find it in the environment as well.
Since it was not defined in the environment scope, the pointer was updated to be
NULL, and only after it meant to be freed. The progarm crashed, becaues the
program attempted to free a NULL pointer. This is now fixed by dealing with the
2 scopes under different vairables. That is, `shell_ptr` and `env_ptr`.