Bug fix: string overflow issue
When the tokenizer loops until the end of a string, the `end_ptr` points
to the null terminator. However, this pointer is incremented further for
the next loop, and may point to an address that should not be accessed
by the function and hence tokenizes characters that don't come from the
prompt.

This commit fixes this issue by always returning a non-null terminator,
that is the last character of a string, right before the null
terminator.